[Update] update the button font
